pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents", "src/db_multiagent_system", "src/config", "src/graph", "src/memory", "src/mcp_server", "src/utils"]

[project]
name = "db-multiagent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.7",
    "mcp>=1.27.0",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [
    ".git",
    ".agents",
    ".venv",
    ".ruff_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
]

[tool.ruff.lint.isort]
known-first-party = ["agents", "config", "db_multiagent_system", "graph", "memory", "mcp_server", "tests", "utils"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: needs Postgres (see README)",
]

[tool.pyrefly]
python-version = "3.12"
search-path = ["src", "."]
project-includes = ["src/**/*.py", "tests/**/*.py"]
